requirements.txt
pandas
streamlit
matplotlib
google-cloud-vision
google-cloud-language
google-api-python-client
Pillow